# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  constants.BudgetCategory:
    enum:
    - Design
    - Material
    - Labor
    - Furniture
    - Appliance
    - Contingency
    - Other
    type: string
    x-enum-varnames:
    - BudgetCategoryDesign
    - BudgetCategoryMaterial
    - BudgetCategoryLabor
    - BudgetCategoryFurniture
    - BudgetCategoryAppliance
    - BudgetCategoryContingency
    - BudgetCategoryOther
...
  constants.SupplierCategory:
    enum:
    - Material
    - Furniture
    - Appliance
    - Labor
    - Design
    - Other
    type: string
    x-enum-varnames:
    - SupplierCategoryMaterial
    - SupplierCategoryFurniture
    - SupplierCategoryAppliance
    - SupplierCategoryLabor
    - SupplierCategoryDesign
    - SupplierCategoryOther
  constants.SupplierStatus:
    enum:
    - Active
    - Suspended
    - Blacklisted
    type: string
    x-enum-varnames:
    - SupplierStatusActive
    - SupplierStatusSuspended
    - SupplierStatusBlacklisted
...
  dto.CreateItemRequest:
    properties:
      budget_amount:
        type: number
      category:
        $ref: '#/definitions/constants.BudgetCategory'
      remark:
        maxLength: 512
        type: string
      sort_order:
        minimum: 0
        type: integer
      sub_category:
        maxLength: 128
        type: string
    required:
    - budget_amount
    - category
    type: object
  dto.CreateReconciliationRequest:
    properties:
      paid_amount:
        minimum: 0
        type: number
      payable_amount:
        type: number
      period:
        type: string
      project_id:
        type: string
      supplier_id:
        type: integer
    required:
    - payable_amount
    - period
    - project_id
    - supplier_id
    type: object
  dto.CreateSupplierRequest:
    properties:
      address:
        maxLength: 255
        type: string
      bank_account:
        maxLength: 128
        type: string
      bank_name:
        maxLength: 128
        type: string
      category:
        $ref: '#/definitions/constants.SupplierCategory'
      contact:
        maxLength: 64
        type: string
      name:
        maxLength: 128
        type: string
      phone:
        maxLength: 32
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
      status:
        $ref: '#/definitions/constants.SupplierStatus'
    required:
    - category
    - name
    type: object
...
  dto.UpdateItemRequest:
    properties:
      budget_amount:
        type: number
      category:
        $ref: '#/definitions/constants.BudgetCategory'
      remark:
        maxLength: 512
        type: string
      sort_order:
        minimum: 0
        type: integer
      sub_category:
        maxLength: 128
        type: string
    type: object
  dto.UpdateReconciliationRequest:
    properties:
      paid_amount:
        minimum: 0
        type: number
      payable_amount:
        type: number
    type: object
  dto.UpdateSupplierRequest:
    properties:
      address:
        maxLength: 255
        type: string
      bank_account:
        maxLength: 128
        type: string
      bank_name:
        maxLength: 128
        type: string
      category:
        $ref: '#/definitions/constants.SupplierCategory'
      contact:
        maxLength: 64
        type: string
      name:
        maxLength: 128
        type: string
      phone:
        maxLength: 32
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
      status:
        $ref: '#/definitions/constants.SupplierStatus'
    type: object
...
  response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:19306
info:
  contact:
    email: support@example.com
    name: API Support
  description: 装修预算与支出管理 API 服务，提供预算编制、支出审批、供应商管理与对账结算能力。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: Renovation Budget API
  version: 1.0.0
paths:
...
  /budgets/{id}/items:
    get:
      parameters:
      - description: 预算表ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 查询预算项列表
      tags:
      - budget-items
    post:
      consumes:
      - application/json
      parameters:
      - description: 预算表ID
        in: path
        name: id
        required: true
        type: integer
      - description: 创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 创建预算项
      tags:
      - budget-items
  /budgets/{id}/items/{item_id}:
    delete:
      parameters:
      - description: 预算表ID
        in: path
        name: id
        required: true
        type: integer
      - description: 预算项ID
        in: path
        name: item_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 删除预算项
      tags:
      - budget-items
    put:
      consumes:
      - application/json
      parameters:
      - description: 预算表ID
        in: path
        name: id
        required: true
        type: integer
      - description: 预算项ID
        in: path
        name: item_id
        required: true
        type: integer
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 更新预算项
      tags:
      - budget-items
...
  /reconciliations:
    get:
      parameters:
      - description: 项目ID
        in: query
        name: project_id
        type: string
      - description: 对账期间
        in: query
        name: period
        type: string
      - description: 供应商ID
        in: query
        name: supplier_id
        type: integer
      - description: 状态
        in: query
        name: status
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 查询对账单列表
      tags:
      - reconciliations
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateReconciliationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 创建对账单
      tags:
      - reconciliations
  /reconciliations/{id}:
    delete:
      parameters:
      - description: 对账单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 删除对账单
      tags:
      - reconciliations
    get:
      parameters:
      - description: 对账单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 获取对账单详情
      tags:
      - reconciliations
    put:
      consumes:
      - application/json
      parameters:
      - description: 对账单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateReconciliationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 更新对账单金额
      tags:
      - reconciliations
  /reconciliations/{id}/confirm:
    post:
      parameters:
      - description: 对账单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 确认对账
      tags:
      - reconciliations
  /reconciliations/{id}/dispute:
    post:
      parameters:
      - description: 对账单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 标记对账争议
      tags:
      - reconciliations
  /reconciliations/{id}/resolve:
    post:
      parameters:
      - description: 对账单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 解决对账争议
      tags:
      - reconciliations
  /suppliers:
    get:
      parameters:
      - description: 合作状态
        in: query
        name: status
        type: string
      - description: 类别
        in: query
        name: category
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 查询供应商列表
      tags:
      - suppliers
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSupplierRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 创建供应商
      tags:
      - suppliers
  /suppliers/{id}:
    delete:
      parameters:
      - description: 供应商ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 删除供应商
      tags:
      - suppliers
    get:
      parameters:
      - description: 供应商ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 获取供应商详情
      tags:
      - suppliers
    put:
      consumes:
      - application/json
      parameters:
      - description: 供应商ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSupplierRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 更新供应商
      tags:
      - suppliers
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"